+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 __BEGIN_UNRELEASED__
 ## [Unreleased]
 ### Added
+- Added `getUserId()` method for fetching the current Heap User ID from the underlying native SDK.
+
 ### Changed
 ### Deprecated
 ### Removed